fix popup UI
When isStripped is true, calendar need self-made border (in img format).
Perhaps this is not what we want. So I change it back to false.

However, when isStripped is false, I currently found no way to insert
spaces between left border/outline and left most calendar date item.
So current walkaround is to also remove spaces in the right side.